Weapon reload rate is respected; weapons don't fire every round of combat if rate > 1. They also fire infinitely fast if rate = 0, and fractional rates are now supported. E.g. with rate 1.5, the weapon will alternate between firing once and twice in a round. We should totally add mounts that affect reload rate...

+++ FrEee/Game/Objects/Technology/WeaponInfo.cs
@@ -36,7 +36,7 @@ namespace FrEee.Game.Objects.Technology
 		/// <summary>
 		/// The number of turns after firing that the weapon must wait to fire again. (1 means the weapon can fire every turn.)
 		/// </summary>
-		public int ReloadRate { get; set; }
+		public double ReloadRate { get; set; }
 
 		/// <summary>
 		/// The display effect used for this weapons.
